+++ packages/shader/src/wgsl/ast.ts
@@ -67,18 +67,10 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
     );
   }
 
-  const skipComments = (nodes: SyntaxNode[]) => nodes.filter(n => n.type.name !== 'Comment');
-
   const getNodes = (node: SyntaxNode, min?: number) => {
     const nodes = getChildNodes(node);
     for (const n of nodes) if (node.type.isError) throwError('error', node);
-    if (min != null && nodes.length < min) throwError(`not enough tokens (${nodes.length} / ${min})`, node);
-    return skipComments(nodes);
-  }
-
-  const getNodesWithAttributes = (node: SyntaxNode, min?: number): [SyntaxNode | null, ...SyntaxNode[]] => {
-    const nodes = getNodes(node, min) as any;
-    if (nodes[0] && nodes[0].type.id !== T.AttributeList) nodes.unshift(null);
+    if (min != null && nodes.length < min) throwError(`not enough tokens (${min})`, node);
     return nodes;
   }
 
@@ -134,9 +126,9 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   };
 
   const getParameter = (node: SyntaxNode): ParameterRef => {
-    const [a, b, c] = getNodesWithAttributes(node, 3);
+    const [a, b, c] = getNodes(node, 3);
 
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const name = getText(b);
     const type = getType(c);
 
@@ -186,16 +178,16 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   };
 
   const getFunction = (node: SyntaxNode): FunctionRef => {
-    const [a, b, c] = getNodesWithAttributes(node, 2);
+    const [a, b, c] = getNodes(node, 2);
 
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const header = getFunctionHeader(b);
 
     const inferred = getInferred(header);
     const {name, type, parameters} = header;
 
     const exclude = parameters ? parameters.map(p => (p as any).name) : undefined;
-    const identifiers = c ? getIdentifiers(c, name, exclude) : NO_STRINGS;
+    const identifiers = c ? getIdentifiers(c, name, exclude) : undefined;
 
     return {name, type, attributes, parameters, identifiers, inferred};
   };
@@ -225,10 +217,10 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   }
 
   const getVariable = (node: SyntaxNode): VariableRef => {
-    const [a, b,, c] = getNodesWithAttributes(node, 2);
+    const [a, b,, c] = getNodes(node, 2);
     const hasValue = !!c;
 
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const {name, type, qual} = getVariableDeclaration(b);
     const value = hasValue ? getText(c) : undefined; 
 
@@ -239,9 +231,11 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   };
 
   const getConstant = (node: SyntaxNode): VariableRef => {
-    const [a, b, c,, d] = getNodesWithAttributes(node, 2);
-
-    const attributes = a ? getAttributes(a) : undefined;
+    const nodes = getNodes(node, 2);
+    
+    const [a, b, c,, d] = nodes;
+    const hasAttributes = a.type.id === T.AttributeList;
+    const attributes = hasAttributes ? getAttributes(a) : undefined;
 
     const hasValue = !!d;
     const {name, type} = getVariableIdentifier(c);
@@ -254,9 +248,9 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   };
   
   const getTypeAlias = (node: SyntaxNode): TypeAliasRef => {
-    const [a,, b,, c] = getNodesWithAttributes(node, 3);
+    const [a,, b,, c] = getNodes(node, 3);
 
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const name = getText(b);
     const type = c ? getType(c) : {name};
 
@@ -264,22 +258,21 @@ export const makeASTParser = (code: string, tree: Tree, name?: string) => {
   };
 
   const getStructMember = (node: SyntaxNode): StructMemberRef => {
-    const [a, b, c] = getNodesWithAttributes(node, 3);
+    const [a, b, c] = getNodes(node, 3);
 
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const name = getText(b);
     const type = getType(c);
 
     return {name, type, attributes};
   };
 
-  const getStructMembers = (node: SyntaxNode): StructMemberRef[] =>
-    getNodes(node).filter(n => getNodes(n).length).map(getStructMember);
+  const getStructMembers = (node: SyntaxNode): StructMemberRef[] => getNodes(node).map(getStructMember);
 
   const getStruct = (node: SyntaxNode): StructRef => {
-    const [a,, b, c] = getNodesWithAttributes(node, 3);
+    const [a,, b, c] = getNodes(node, 3);
     
-    const attributes = a ? getAttributes(a) : undefined;
+    const attributes = getAttributes(a);
     const name = getText(b);
     const members = getStructMembers(c);